predict/__init__.py
import logging
import json
import azure.functions as func

# Import helper script
from .predict import predict_image_from_url

app = func.FunctionApp()


def main(req: func.HttpRequest) -> func.HttpResponse:
    image_url = req.params.get('img')
    logging.info('Image URL received: ' + image_url)

    results = predict_image_from_url(image_url)

    headers = {
        "Content-type": "application/json",
        "Access-Control-Allow-Origin": "*"
    }

    return func.HttpResponse(json.dumps(results), headers=headers)
