requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
requests
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.13.0+cpu
torchvision==0.14.0+cpu
black # lint & formatting helper
flake8  # lint & formatting helper
isort  # lint & formatting helper
sphinx  # documentation
sphinx-autoapi  # documentation
sphinx-mdinclude  # documentation
sphinx-rtd-theme  # read-the-docs theme
docstr-coverage  # doc string coverage
wily
coverage
coverage-badge
docstr-coverage
pybadges
pylint
pytest